requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
Werkzeug==2.1.2